FB-15a: extract prompt metadata artifact writer to TaskPromptArtifacts.ts
- Moves the ~120-line writePromptMetadataArtifacts implementation out of
  src/core/task/index.ts into a side-effect-isolated helper module.
- Task.writePromptMetadataArtifacts becomes a thin façade; DiracContext
  receives an arrow wrapper so the public callback shape is preserved.
- Verified: core/task tests 283 passing, tsc baseline unchanged, biome clean.
