+++ src/core/task/index.ts
@@ -73,7 +73,6 @@ import { type Mode } from "@shared/storage/types"
 import { isMutatingTool } from "@shared/tools"
 import { DiracAskResponse } from "@shared/WebviewMessage"
 import { isLocalModel, isParallelToolCallingEnabled } from "@utils/model-utils"
-import fs from "fs/promises"
 import Mutex from "p-mutex"
 import pWaitFor from "p-wait-for"
 import * as path from "path"
@@ -101,6 +100,7 @@ import { StreamingMetricsManager } from "./StreamingMetricsManager"
 import { StreamResponseHandler } from "./StreamResponseHandler"
 import { type SteeringClaim } from "./steering"
 import { TaskMessenger } from "./TaskMessenger"
+import { type TaskPromptArtifactsContext, writePromptMetadataArtifacts } from "./TaskPromptArtifacts"
 import { TaskState } from "./TaskState"
 import {
 	appendQueuedSteeringToNextApiRequest,
@@ -179,6 +179,14 @@ export class Task {
 			taskMessenger: this.taskMessenger,
 			postStateToWebview: () => this.postStateToWebview(),
 			withStateLock: (fn) => this.withStateLock(fn),
+		}
+	}
+
+	private get promptArtifactsContext(): TaskPromptArtifactsContext {
+		return {
+			taskId: this.taskId,
+			cwd: this.cwd,
+			stateManager: this.stateManager,
 		}
 	}
 
@@ -707,7 +715,7 @@ export class Task {
 			getCurrentProviderInfo: this.getCurrentProviderInfo.bind(this),
 			getEnvironmentDetails: this.getEnvironmentDetails.bind(this),
 			getPinnedContext: () => this.taskState.pinnedContext,
-			writePromptMetadataArtifacts: this.writePromptMetadataArtifacts.bind(this),
+			writePromptMetadataArtifacts: (params) => writePromptMetadataArtifacts(this.promptArtifactsContext, params),
 			handleHookCancellation: this.hookManager.handleHookCancellation.bind(this.hookManager),
 			setActiveHookExecution: this.hookManager.setActiveHookExecution.bind(this.hookManager),
 			clearActiveHookExecution: this.hookManager.clearActiveHookExecution.bind(this.hookManager),
@@ -1140,115 +1148,7 @@ export class Task {
 		fullHistory?: any[]
 		deletedRange?: [number, number]
 	}): Promise<void> {
-		const enabledSetting = this.stateManager.getGlobalSettingsKey("writePromptMetadataEnabled")
-		const enabledFlag = process.env.DIRAC_WRITE_PROMPT_ARTIFACTS?.toLowerCase()
-		const enabled =
-			enabledSetting ||
-			enabledFlag === "1" ||
-			enabledFlag === "true" ||
-			enabledFlag === "yes" ||
-			process.env.IS_DEV === "true"
-		if (!enabled) {
-			return
-		}
-
-		try {
-			// Env var is OS-level (user-controlled, safe to allow absolute); workspace setting is the exfiltration vector.
-			const envDir = process.env.DIRAC_PROMPT_ARTIFACT_DIR?.trim()
-			const settingDir = this.stateManager.getGlobalSettingsKey("writePromptMetadataDirectory")?.trim()
-			const cwdResolved = path.resolve(this.cwd)
-			// Setting-configured dirs must resolve under cwd to prevent workspace settings from exfiltrating prompts.
-			// Only validate the setting when no env var is provided — env takes precedence and is trusted.
-			if (!envDir && settingDir) {
-				const resolved = path.isAbsolute(settingDir) ? path.resolve(settingDir) : path.resolve(this.cwd, settingDir)
-				if (resolved !== cwdResolved && !resolved.startsWith(cwdResolved + path.sep)) {
-					Logger.warn(`[Task ${this.taskId}] writePromptMetadataDirectory outside cwd rejected: ${resolved}`)
-					return
-				}
-			}
-			const configuredDir = envDir || settingDir
-			const artifactDir = configuredDir
-				? path.isAbsolute(configuredDir)
-					? path.resolve(configuredDir)
-					: path.resolve(this.cwd, configuredDir)
-				: path.resolve(this.cwd, ".dirac-prompt-artifacts")
-
-			await fs.mkdir(artifactDir, { recursive: true })
-			// Defense-in-depth: re-check the boundary after mkdir resolves any symlinks in the path,
-			// so a workspace-planted symlink can't exfiltrate prompts outside cwd.
-			// Only enforced for setting-derived paths — env vars are OS-level and may legitimately point anywhere.
-			let writeDir = artifactDir
-			if (!envDir) {
-				const realArtifactDir = await fs.realpath(artifactDir)
-				if (realArtifactDir !== cwdResolved && !realArtifactDir.startsWith(cwdResolved + path.sep)) {
-					Logger.warn(
-						`[Task ${this.taskId}] artifact dir resolves outside cwd (symlink?), rejected: ${realArtifactDir}`,
-					)
-					return
-				}
-				writeDir = realArtifactDir
-			}
-			// Ensure the artifact dir is git-ignored so debug dumps don't get committed.
-			const gitignorePath = path.join(writeDir, ".gitignore")
-			await fs.writeFile(gitignorePath, "*\n!.gitignore\n", "utf8").catch(() => {})
-
-			const debugPath = path.join(writeDir, `task-${this.taskId}-debug.md`)
-
-			let markdown = `## System Prompt\n\n${params.systemPrompt}\n\n`
-
-			if (params.tools) {
-				markdown += `## Tools\n\n\`\`\`json\n${JSON.stringify(params.tools, null, 2)}\n\`\`\`\n\n`
-			}
-
-			if (params.fullHistory) {
-				markdown += `## Conversation History\n\n`
-				const [deletedStart, deletedEnd] = params.deletedRange || [-1, -1]
-
-				for (let i = 0; i < params.fullHistory.length; i++) {
-					const message = params.fullHistory[i]
-					const isTruncated = i >= deletedStart && i <= deletedEnd
-
-					markdown += `### [${message.role.toUpperCase()}]${isTruncated ? " [TRUNCATED]" : ""}\n`
-
-					if (typeof message.content === "string") {
-						markdown += `${message.content}\n\n`
-					} else if (Array.isArray(message.content)) {
-						for (const block of message.content) {
-							if (block.type === "text") {
-								markdown += `**Text:** ${block.call_id ? `(\`call_id: ${block.call_id}\`)` : ""}\n${block.text}\n\n`
-							} else if (block.type === "thinking") {
-								markdown += `**Thinking:** ${block.call_id ? `(\`call_id: ${block.call_id}\`)` : ""}\n${block.thinking}\n\n`
-							} else if (block.type === "redacted_thinking") {
-								markdown += `**Thinking:** [Redacted] ${block.call_id ? `(\`call_id: ${block.call_id}\`)` : ""}\n\n`
-							} else if (block.type === "tool_use") {
-								markdown += `**Tool Use:** \`${block.name}\` (\`id: ${block.id}\`, \`call_id: ${block.call_id}\`)\n`
-								markdown += `\`\`\`json\n${JSON.stringify(block.input, null, 2)}\n\`\`\`\n\n`
-							} else if (block.type === "tool_result") {
-								markdown += `**Tool Result:** (\`${block.tool_use_id}\`)\n`
-								if (typeof block.content === "string") {
-									markdown += `${block.content}\n\n`
-								} else if (Array.isArray(block.content)) {
-									for (const contentBlock of block.content) {
-										if (contentBlock.type === "text") {
-											markdown += `${contentBlock.text}\n\n`
-										} else if (contentBlock.type === "image") {
-											markdown += `[Image: ${contentBlock.source?.type}]\n\n`
-										}
-									}
-								}
-							} else if (block.type === "image") {
-								markdown += `[Image: ${block.source?.type}]\n\n`
-							}
-						}
-					}
-					markdown += "---\n\n"
-				}
-			}
-
-			await fs.writeFile(debugPath, markdown, "utf8")
-		} catch (error) {
-			Logger.error("Failed to write prompt metadata artifacts:", error)
-		}
+		return writePromptMetadataArtifacts(this.promptArtifactsContext, params)
 	}
 
 	private getApiRequestIdSafe(): string | undefined {
@@ -1419,7 +1319,7 @@ export class Task {
 			this.stateManager.getGlobalSettingsKey("useAutoCondense"),
 		)
 
-		await this.writePromptMetadataArtifacts({
+		await writePromptMetadataArtifacts(this.promptArtifactsContext, {
 			systemPrompt,
 			providerInfo,
 			tools: toolSnapshot.nativeTools,
